extract layoutqueue to separate file
